# src/backend/libs/efiling-bambora-api-client/payments-swagger.yaml
swagger: "2.0"
info:
  title: Bambora Payment APIs
  description: "https://api.na.bambora.com/v1"
  version: 1.0.5
host: api.na.bambora.com
schemes:
  - https
basePath: /v1
consumes:
  - application/json
produces:
  - application/json
securityDefinitions:
  Passcode:
    type: apiKey
    in: header
    name: Authorization
    description: 'Pass the authorization header in your request as "Authorization: Passcode Base64Encoded(merchant_id:passcode)"'
  Sub-Merchant-Id:
    type: apiKey
    in: header
    name: Sub-Merchant-Id
    description: 'OPTIONAL: Only used by partners to make an API call on behalf of their subMerchant. Pass the header in your request as "Sub-Merchant-Id: sub_merchant_id"'
security:
  - Passcode: []
  - Sub-Merchant-Id: []
paths:
...
  /reports:
    post:
      summary: Search Query
      description: "Query for transactions using a date range and optional search criteria. This method allows you to page your search results if you are expecting a lot of results to be returned. The page start value begins at 1. If no records are found the API will return a 200 response message with an empty array. For details on the parameters and allowed values for Criteria please visit the documentation http://dev.na.bambora.com"
      tags:
        - Reporting
      operationId: getReport
      parameters:
        - name: search_query
          in: body
          schema:
            $ref: "#/definitions/SearchQuery"
      responses:
        200:
          description: A transaction object.
          schema:
            $ref: "#/definitions/SearchResult"
        400:
          $ref: "#/responses/BadRequest"
        401:
          $ref: "#/responses/Unauthorized"
        402:
          $ref: "#/responses/PaymentRequired"
        403:
          $ref: "#/responses/Forbidden"
        405:
          $ref: "#/responses/MethodNotAllowed"
        500:
          $ref: "#/responses/InternalServerError"
...
responses:
  BadRequest:
    description: Bad Request
    schema:
      $ref: "#/definitions/ErrorResponse"
  Unauthorized:
    description: Authentication Failure
    schema:
      $ref: "#/definitions/ErrorResponse"
  PaymentRequired:
    description: Business Rule Violation or Decline
    schema:
      $ref: "#/definitions/ErrorResponse"
  Forbidden:
    description: Authorization Failure
    schema:
      $ref: "#/definitions/ErrorResponse"
  MethodNotAllowed:
    description: Invalid Request Method
    schema:
      $ref: "#/definitions/ErrorResponse"
  InternalServerError:
    description: Internal Server Error
    schema:
      $ref: "#/definitions/ErrorResponse"
definitions:
...
  SearchQuery:
    required:
      - name
      - start_date
      - end_date
      - start_row
      - end_row
    properties:
      name:
        type: string
        description: "Can be either 'Search' for all fields or 'TransHistoryMinimal' for a subset of the fields returned in the results."
        enum: ["Search", "TransHistoryMinimal"]
      start_date:
        type: string
        description: "The start date (inclusive) '2015-04-22T10:03:19' in the timezone of your merchant account."
      end_date:
        type: string
        description: "The end date (inclusive) '2015-04-22T10:03:19' in the timezone of your merchant account."
      start_row:
        type: number
        format: integer
        description: Used to page the results. 1-based
        default: 1
      end_row:
        type: number
        format: integer
        description: Used to page the results. 1-based. This should always be 1 larger than start_row.
        default: 2
      criteria:
        description: Optional search criteria. All criteria are ANDed together.
        type: array
        items:
          $ref: "#/definitions/Criteria"
  Criteria:
    properties:
      field:
        type: number
        format: integer
        description: "Refer to the table on this page for the searchable fields http://dev.na.bambora.com"
      operator:
        type: string
        description: "URL encoded comparators such as less than, greater than, equals... Refer to the table on this page for the full list of comparators http://dev.na.bambora.com"
        enum: ["%3D", "%3C", "%3E", "%3C%3D", "%3E%3D", "START%20WITH"]
      value:
        type: string
        description: The value you want to match against.
  SearchResult:
    properties:
      records:
        type: array
        items:
          $ref: "#/definitions/SearchRecord"
  SearchRecord:
    properties:
      row_id:
        type: number
        format: integer
      trn_id:
        type: number
        format: integer
      trn_date_time:
        type: string
        description: "2015-04-22T10:03:19.323-07:00"
      trn_type:
        type: string
      trn_order_number:
        type: string
      trn_payment_method:
        type: string
      trn_comments:
        type: string
      trn_masked_card:
        type: string
        description: "The credit card with the middle digits redacted with X's"
      trn_amount:
        type: number
        format: float32
      trn_returns:
        type: number
        format: float32
      trn_completions:
        type: number
        format: float32
      trn_voided:
        type: number
        format: integer
      trn_response:
        type: number
        format: integer
      trn_card_type:
        type: string
        description: MC VI etc
      trn_batch_no:
        type: number
        format: integer
      trn_avs_result:
        type: string
        description: Address Verification Service
      trn_cvd_result:
        type: number
        format: integer
      trn_cavv_result:
        type: string
        description: 3D Secure Code Result
      trn_card_expiry:
        type: string
      message_id:
        type: number
        format: integer
      message_text:
        type: string
      trn_card_owner:
        type: string
      trn_ip:
        type: string
      trn_approval_code:
        type: string
      trn_reference:
        type: number
        format: integer
      b_name:
        type: string
      b_email:
        type: string
      b_phone:
        type: string
      b_address1:
        type: string
      b_address2:
        type: string
      b_city:
        type: string
      b_province:
        type: string
      b_postal:
        type: string
      b_country:
        type: string
      s_name:
        type: string
      s_email:
        type: string
      s_phone:
        type: string
      s_address1:
        type: string
      s_address2:
        type: string
      s_city:
        type: string
      s_province:
        type: string
      s_postal:
        type: string
      s_country:
        type: string
      ref1:
        type: string
      ref2:
        type: string
      ref3:
        type: string
      ref4:
        type: string
      ref5:
        type: string
      product_name:
        type: string
      product_id:
        type: string
      customer_code:
        type: string
        description: "The Payment Profile ID used in this transaction, if appropriate."
      currency_abbr:
        type: string
        maxLength: 3
        description: "The currency for this transaction."
      merchant_id:
        type: integer
        description: "The id of the merchant for this transaction."
      merchant_name:
        type: string
        description: "The name of the merchant for this transaction."
      entry_method:
        type: string
        maxLength: 1
        description: "How the transaction was initiated. N = Standard Entry, A = Apple Pay, G = Android Pay, M = MasterPass, S = EMV Swipe, V = Visa Checkout, Y = Unimag Card Swiper, T = EMV Chip Tap, E = EMV Chip Dip"
        enum: ["N", "A", "G", "M", "S", "V", "Y", "T", "E"]
      authorizing_merchant_id:
        type: integer
        description: "The id of the merchant that authorized this transaction."
...
  ErrorResponse:
    properties:
      code:
        type: integer
        format: int32
      category:
        type: integer
        format: int32
      message:
        type: string
      reference:
        type: string
        description: Not used. Always returns an empty string.
      details:
        type: array
        items:
          $ref: "#/definitions/Detail"
      validation:
        $ref: "#/definitions/CardValidation"
  CardValidation:
    properties:
      id:
        type: string
        maxLength: 9
        description: Transaction ID
      approved:
        type: integer
        description: "Approval status of payment transaction. 0 if the transaction is not approved. 1 if the transaction is approved."
      message_id:
        type: integer
        maxLength: 3
        description: "Payment response code"
      message:
        type: string
        maxLength: 256
        description: "Message containing information about the transactions status"
      auth_code:
        type: string
        maxLength: 32
        description: "Auth Code"
      trans_date:
        type: string
        description: "The date it was created '2015-04-22T10:03:19.323-07:00'"
      order_number:
        type: string
        maxLength: 32
        default: "Transaction Id"
        description: "Order number"
      type:
        type: string
        description: "Payment transaction type"
        enum: ["P", "R", "VP", "VR", "PA", "PAC"]
      amount:
        type: number
        format: double
        description: "A decimal value in dollars. Uses up to two decimal places."
      cvd_id:
        type: integer
        maxLength: 2
        description: "CVD Response ID"
  Detail:
    properties:
      field:
        type: string
      message:
        type: string
...